List notable implementation details of the various engines. (#59)
* List notable implementation details of the various engines.

It would be interesting for readers of the benchmark results to better
understand how the various engines differ. While many readers stop at Rust vs.
Java vs. C++, there are lots of other notable implementation differences such
as WAND vs. MAXSCORE or whether or not doc ID reordering is used.

I thought it was also interesting to highlight similarities, such as the use of
mmap and PFOR-delta.

* Make implementation details hidden by default.

* Undo unintended changes

* Tantivy uses FOR, not PFOR.

* Update engines/tantivy-0.22/details.json

Co-authored-by: Paul Masurel <paul@quickwit.io>

* Update engines/pisa-0.8.2/details.json

Co-authored-by: Paul Masurel <paul@quickwit.io>

---------

Co-authored-by: Paul Masurel <paul@quickwit.io>